+++ CHANGELOG.md
@@ -7,8 +7,57 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Table of Contents
 
+- [1.0.0 - 2021-09-30 - First release](#100---2021-09-30---first-release)
+- [0.1.4 - 2021-09-29 - Game logic finish](#014---2021-09-29---game-logic-finish)
+- [0.1.3 - 2021-09-28 - Game logic setup](#013---2021-09-28---game-logic-setup)
+- [0.1.2 - 2021-09-27 - Base page setup](#012---2021-09-27---base-page-setup)
 - [0.1.1 - 2021-09-21 - Initial release](#011---2021-09-21---initial-release)
 - [0.1.0 - 2021-09-21 - Pre release](#010---2021-09-21---pre-release)
+
+## [1.0.0] - 2021-09-30 - First release
+
+### Added
+
+- Update this `CHANGELOG`
+- Release on [netlify.com](https://www.netlify.com)
+
+## [0.1.4] - 2021-09-29 - Game logic finish
+
+### Added
+
+- Score calculation to the game
+- Card flipping logic
+- `WinningScreenModal` to display an end game window
+
+### Fixed
+
+- `SettingsModal` minor error
+
+## [0.1.3] - 2021-09-28 - Game logic setup
+
+### Added
+
+- `Game` component to hold game logic
+- `Card` component for visual representation of a card
+
+## [0.1.2] - 2021-09-27 - Base page setup
+
+### Added
+
+- `Navbar` component
+- A `modal` for setting the game difficulty
+- Images for each card pair
+- A `footer` section
+- A `Hero` component for the front page
+
+### Removed
+
+- `reporWebVitals.js` because unnecessary
+- `App.css` instead use `index.css`
+
+### Fixed
+
+- Site title
 
 ## [0.1.1] - 2021-09-21 - Initial release
 
@@ -30,5 +79,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - An [MIT](https://opensource.org/licenses/MIT) `LICENSE` for a basic general purpose license
 - `.gitignore` to exclude unnecesarry files from versioning
 
-[0.1.1]: https://github.com/gaborkristof/react-memory
-[0.1.0]: https://github.com/gaborkristof/react-memory
+[1.0.0]: https://github.com/gaborkristof/react-memory
+[0.1.4]: https://github.com/grabovszky/react-memory/tree/7c56f5faa02f21b93c42d49b65003ee5ff657322
+[0.1.3]: https://github.com/grabovszky/react-memory/tree/b8bfd0088497022047c586cb474a3b051a851146
+[0.1.2]: https://github.com/grabovszky/react-memory/tree/b8bfd0088497022047c586cb474a3b051a851146
+[0.1.1]: https://github.com/grabovszky/react-memory/tree/69231751adfdb2552b1991a1703365c8b52d175d
+[0.1.0]: https://github.com/grabovszky/react-memory/tree/f303beef0b66ef7fe3cc1bc03c48a8494b8f38c2